Move track improvements to thread
Create vpKwiverImproveTrackTask to encapsulate the actual execution of
the algorithm for improving tracks. In particular, the new class uses a
separate thread for the algorithm execution, and includes support for
progress reporting (which is now connected to a progress dialog in
vpViewCore). This should avoid the application being non-responsive
during execution of the algorithm.